+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Notable changes are documented in this file, whose format follows [Keep a Change
 ### Fixed
 
 - Handle all native test binaries instead of just the first
-- Compose Multiplatform compatibility on Apple platforms (#141)
+- Compose Multiplatform compatibility on iOS (#141)
 
 ## [0.11.0] - 2025-12-13